agent_hub/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
httpx>=0.27.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
openai>=1.40.0
anthropic>=0.34.0
slack_sdk>=3.33.0
requests>=2.32.3
PyPDF2>=3.0.1
filetype>=1.2.0
openpyxl>=3.1.5
